add abstract observable type
